TYPE-A: narrow SpaceGroupOp rotation i32 -> i8
Rotation-matrix entries for crystallographic symmetry operations sit in
{-1, 0, 1} for cubic groups and {-2, -1, 0, 1, 2} at worst for hexagonal
settings. i8 gives ~40x headroom and shaves 27 bytes per op, packing all
48 Fd-3m ops into ~2 cache lines (from 27).

- SymmOp / SpaceGroupOp::rotation: [[i32; 3]; 3] -> [[i8; 3]; 3]
- Arithmetic widens to i32 at method entry (rotation_i32 helper) so
  det/compose/adjugate triple products never risk overflow.
- SymmOp::from_flat retains its [i32; 9] public signature; it narrows via
  i8::try_from to preserve call-site ergonomics.
- detect::narrow_rotation bridges the i32 enumeration path (unchanged) to
  the i8 storage.
- check_grid_compatibility and real-space symmetrize_density widen reads
  via i32::from / f64::from explicitly; these are cold paths.
- symmetrize_density_g's transpose_rotation folds the i8->i32 widen into
  its existing transpose pass, so the hot loop's arithmetic stays i32.

No arithmetic semantics change; all 265+ tests pass bit-identical.

+++ src/symmetry/density/g_space.rs
@@ -24,13 +24,18 @@ fn rotate_miller(r: &[[i32; 3]; 3], n: [i32; 3]) -> [i32; 3] {
     ]
 }
 
-/// Transpose of an integer 3×3 rotation matrix.
+/// Transpose of an integer 3×3 rotation matrix, widened to `i32`.
+///
+/// `SpaceGroupOp` stores rotations as `i8` for cache-density (per TYPE-A);
+/// the inner symmetrizer loop keeps multiplication in `i32` to preserve
+/// overflow-free behavior for `R^T · n` where `|n_i| ≤ N/2`. This function
+/// performs both the transpose and the `i8 → i32` widen in a single pass.
 #[inline]
-fn transpose_rotation(r: &[[i32; 3]; 3]) -> [[i32; 3]; 3] {
+fn transpose_rotation(r: &[[i8; 3]; 3]) -> [[i32; 3]; 3] {
     [
-        [r[0][0], r[1][0], r[2][0]],
-        [r[0][1], r[1][1], r[2][1]],
-        [r[0][2], r[1][2], r[2][2]],
+        [i32::from(r[0][0]), i32::from(r[1][0]), i32::from(r[2][0])],
+        [i32::from(r[0][1]), i32::from(r[1][1]), i32::from(r[2][1])],
+        [i32::from(r[0][2]), i32::from(r[1][2]), i32::from(r[2][2])],
     ]
 }
